% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_metadata.R
\name{collect_metadata}
\alias{collect_metadata}
\title{Read and combine raw metadata}
\usage{
collect_metadata(df_structure, roxas_version)
}
\arguments{
\item{df_structure}{Dataframe with all input filenames}

\item{roxas_version}{ROXAS version (used to read the settings files)}
}
\value{
A dataframe containing the extracted data.
}
\description{
Collect the ROXAS settings and image exif data from all raw files and
combine into one dataframe with the data structure.
}
