% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs2xcell-package.R
\docType{package}
\name{rxs2xcell-package}
\alias{rxs2xcell}
\alias{rxs2xcell-package}
\title{rxs2xcell: Preprocess QWA Data (from ROXAS) and Upload to XCELL Database}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mara Naegelin \email{mara.naegelin@wsl.ch} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
