% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_metadata.R
\name{subset_treecodes}
\alias{subset_treecodes}
\title{Include or exclude tree codes from data collection}
\usage{
subset_treecodes(df_structure, include_codes = NULL, exclude_codes = NULL)
}
\arguments{
\item{df_structure}{The dataframe containing all input filenames and data structure.}

\item{include_codes}{EITHER provide a vector of the tree codes to be included}

\item{exclude_codes}{OR provide a vector of the tree codes to be excluded}
}
\value{
df_structure with the filtered tree codes.
}
\description{
This function filters the data structure dataframe based on the tree codes provided
}
