% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{validate_QWA_data}
\alias{validate_QWA_data}
\title{Validate the raw QWA data}
\usage{
validate_QWA_data(QWA_data)
}
\value{
validated QWA_data.
}
\description{
Initial checks to ensure the quality of the raw QWA data.
This function checks for the following issues:
\itemize{
\item undated images (i.e. YEAR is NA or in the future)
\item images without cell wall thickness estimates
If any images with with these issues are found, the function will stop and
ask the user to correct the raw data before proceeding.
}
}
\details{
Next, the function identifies the rings with the following issues:
\itemize{
\item incomplete rings
\item missing rings
\item duplicate rings
}

Here, by incomplete rings we mean those at the inner (pith) and outer (bark)
borders of an images, where some cells are recognized but a MRW can NOT
(outer) or NOT ACCURATELY (inner) be estimated.
NOTE: Because ROXAS uses the outer ring boundary to estimate MRW, the
innermost ring generally has an MRW estimate (that is not based on the true
ring boundary but rather the image border), while the outermost ring has no
MRW value (except if it is actually complete either because it is at the
at the bark or because the user removed the incomplete ring manually in ROXAS).
Therefore, we always flag the innermost ring  per image as incomplete,
while the outermost ring is incomplete if and only if it has no MRW.
TODO: what if the user has manually removed the outermost ring in ROXAS?

Missing rings are for years that have no discernible ring in the image, but
have been manually added in ROXAS during cross-dating, leading to an entry
in the rings data but no corresponding no entries (cells) in the cells data.

Duplicate rings are those that are present in multiple images due to them
overlapping. All years which have cells in more than one image are flagged,
with the exception of the (complete) year with the highest number of cells
for each overlap, which is the one that would usually be selected for further
analysis when building chronologies.
}
