% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs2xcell-package.R
\docType{package}
\name{rxs2xcell-package}
\alias{rxs2xcell}
\alias{rxs2xcell-package}
\title{rxs2xcell: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
